+++ src/votekit/elections/election_types/ranking/borda.py
@@ -7,7 +7,7 @@ from ....utils import (
     validate_score_vector,
     score_profile_from_rankings,
 )
-from typing import Optional, Union, Sequence
+from typing import Optional, Union, Sequence, Literal
 from fractions import Fraction
 from functools import partial
 
@@ -16,9 +16,9 @@ class Borda(RankingElection):
     r"""
     Borda election. Positional voting system that assigns a decreasing number of points to
     candidates based on their ordering. The conventional score vector is :math:`(n, n-1, \dots, 1)`
-    where :math:`n` is the number of candidates. Candidates with the highest scores are elected.
-    This class uses the `utils.score_profile_from_rankings()` to handle ballots with ties and
-    missing candidates.
+    where :math:`n` is the ``max_ballot_length`` of the profile. Candidates with the highest scores
+    are elected. This class uses the ``utils.score_profile_from_rankings()`` to handle ballots with
+    ties.
 
     Args:
         profile (PreferenceProfile): Profile to conduct election on.
@@ -28,6 +28,12 @@ class Borda(RankingElection):
             If it is shorter, we add 0s. Defaults to None, which is the conventional Borda vector.
         tiebreak (str, optional): Tiebreak method to use. Options are None, 'random', and
             'first_place'. Defaults to None, in which case a tie raises a ValueError.
+        scoring_tie_convention (Literal["high", "average", "low"], optional): How to award points
+            for tied rankings. Defaults to "low", where any candidates tied receive the lowest
+            possible points for their position, eg three people tied for 3rd would each receive the
+            points for 5th. "high" awards the highest possible points, so in the previous example,
+            they would each receive the points for 3rd. "average" averages the points, so they would
+            each receive the points for 4th place.
 
     """
 
@@ -37,16 +43,20 @@ class Borda(RankingElection):
         m: int = 1,
         score_vector: Optional[Sequence[Union[float, Fraction]]] = None,
         tiebreak: Optional[str] = None,
+        scoring_tie_convention: Literal["high", "average", "low"] = "low",
     ):
         self.m = m
         self.tiebreak = tiebreak
         if not score_vector:
-            score_vector = list(range(len(profile.candidates), 0, -1))
+            score_vector = list(range(profile.max_ballot_length, 0, -1))
 
         validate_score_vector(score_vector)
         self.score_vector = score_vector
         score_function = partial(
-            score_profile_from_rankings, score_vector=score_vector, to_float=False
+            score_profile_from_rankings,
+            score_vector=score_vector,
+            to_float=False,
+            tie_convention=scoring_tie_convention,
         )
         super().__init__(profile, score_function=score_function, sort_high_low=True)